Refactor: Move mouse cursor classes to "services" package (#77751)
