+++ packages/flutter/test/rendering/mouse_tracker_cursor_test.dart
@@ -10,7 +10,7 @@ import 'package:flutter/gestures.dart';
 import 'package:flutter/services.dart';
 
 import '../flutter_test_alternative.dart';
-import './mouse_tracking_test_utils.dart';
+import 'mouse_tracker_test_utils.dart';
 
 typedef MethodCallHandler = Future<dynamic> Function(MethodCall call);
 typedef SimpleAnnotationFinder = Iterable<HitTestTarget> Function(Offset offset);
